% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bupaverse.R
\docType{package}
\name{bupaverse-package}
\alias{bupaverse}
\alias{bupaverse-package}
\title{bupaverse: Easily Install and Load the 'bupaverse'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'bupaverse' is an open-source, integrated suite of R-packages for handling and analysing business process data, developed by the Business Informatics research group at Hasselt University, Belgium. Profoundly inspired by the 'tidyverse' package, the 'bupaverse' package is designed to facilitate the installation and loading of multiple 'bupaverse' packages in a single step. Learn more about 'bupaverse' at the \url{https://bupar.net} homepage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bupar.net/}
  \item \url{https://github.com/bupaverse/bupaverse/}
  \item \url{https://bupaverse.github.io/bupaverse/}
}

}
\author{
\strong{Maintainer}: Gert Janssenswillen \email{gert.janssenswillen@uhasselt.be} (\href{https://orcid.org/0000-0002-7474-2088}{ORCID})

Authors:
\itemize{
  \item Gerard van Hulzen \email{gerard.vanhulzen@uhasselt.be} (\href{https://orcid.org/0000-0001-8962-9515}{ORCID})
}

}
\keyword{internal}
