% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{bupaverse_conflicts}
\alias{bupaverse_conflicts}
\title{Conflicts between the bupaverse and other packages}
\usage{
bupaverse_conflicts()
}
\value{
Returns a \code{\link{list}} all the conflicts between packages in the \code{\link{bupaverse}} and other packages that you have loaded.
}
\description{
This function lists all the conflicts between packages in the \code{\link{bupaverse}} and other packages that you have loaded.
}
\examples{
bupaverse_conflicts()

}
