% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{bupaverse_packages}
\alias{bupaverse_packages}
\title{List all packages in the bupaverse}
\usage{
bupaverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{\code{\link{logical}} (default \code{TRUE}): Include \code{\link{bupaverse}} in the list?}
}
\value{
Returns a \code{\link{list}} of all packages included in the \code{\link{bupaverse}}.
}
\description{
List all packages in the bupaverse
}
\examples{
bupaverse_packages()

}
