% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bupaverse.R
\docType{package}
\name{bupaverse-package}
\alias{bupaverse}
\alias{bupaverse-package}
\title{bupaverse: Easily Install and Load the 'bupaverse'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'bupaverse' is an open-source, integrated suite of R-packages for the handling and analysis of business process data, developed by the Business Informatics research group at Hasselt University, Belgium. This package is designed to facilitate the installation and loading multiple 'bupaverse' packages in a single step. Learn more about 'bupverse' at \url{https://www.bupar.net}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.bupar.net}
  \item \url{https://github.com/bupaverse/bupaverse}
  \item \url{https://bupaverse.github.io/bupaverse/}
}

}
\author{
\strong{Maintainer}: Gerard van Hulzen \email{gerard.vanhulzen@uhasselt.be} (\href{https://orcid.org/0000-0001-8962-9515}{ORCID})

Authors:
\itemize{
  \item Gert Janssenswillen \email{gert.janssenswillen@uhasselt.be} (\href{https://orcid.org/0000-0002-7474-2088}{ORCID})
}

}
\keyword{internal}
